+++ CHANGELOG.md
@@ -6,6 +6,19 @@ More infos about this file : http://keepachangelog.com/
 
 ## [Unreleased] - no_due_date
 
+## [v1.1.7] - 2019-11-20
+
+- **Add a method to `TimeIntervalForDate` to convert between two `timeZone`**
+
+## [v1.1.6] - 2019-10-28
+
+- **Remove `private[core]` on `Schedule`**
+
+## [v1.1.5] - 2019-09-19
+
+Bug fix:
+- **split segments for a time range around midnight crash**
+
 ## [v1.1.1] - 2019-05-16
 
 Bug fix: